Simplify and consolidate rescue/unrescue
When running tempest against nova-lxd, tests for rescue
and unrescue instances failed. This is due to various
API changes in LXD and in nova.

Now we have simplified rescuing an instance, thanks to the
container config refactoring, by doing the
following:

1. Stop the old container.
2. Rename the old contianer to instance.name-backup.
3. Re-use the old nova instnace object of the original
   conainer.
4. Append the instnce.name-backup path as a disk
   configuration.
5. Create a new container and start it.

In order to accomplish this we consolidated the rescue
method, and removed the need for boolean flags to determine
the instance is a rescue container or not.

Signed-off-by: Chuck Short <chuck.short@canonical.com>