Fix config drive
Running tempest tests against nova-lxd discovered LXD is not working
properly.

ISO creation and handling was not working properly. Initially,
it was assumed that users can mount and ISO in their unprivileged
container. However this is not the case, which makes injecting
the ISO of the configdrive into an instance useless.

Now when a configdrive is needed, we create the ISO outside the
container, mount and copy the files to an directory on the host
and inject the path into the container.

Signed-off-by: Chuck Short <chuck.short@canonical.com>