Fix migration with configdrive
Migration fails with configdrive if there is a configdrive attached
to the instance.

After the container is created on the new host via init API call
the LXD daemon expects the directory that is originated to be on
the host to be present on the new host. Otherwise the container
will not start.

This is not tested by tempest and it should probably be included.

Signed-off-by: Chuck Short <chuck.short@canonical.com>