Fix nova-lxd driver loading
Recent nova commit 8eb03de1eb83a6cd2d4d41804e1b8253f94e5400 removed the
mechanism by which nova-lxd was loading its ComputeDriver from out
of tree.

With this change we can now add a nova.virt.lxd
to the nova.virt namespace from the nova-lxd package.
This makes it easier to maintain and easier to merge
into the nova tree when it is ready.

Signed-off-by: Chuck Short <chuck.short@canonical.com>