Add mocking to ensure test isolation from fs
Commit 42803c39867b8b932be61c6e5536e5590932bd8b fixed some issues
around support for config-drive with nova-lxd, introducing some
checks to ensure that the instance directory exists prior to
creating and spawning the associated container.

In an environment where the Python module install location is
read-only, the unit test for spawn instance fails:

 OSError: [Errno 13] Permission denied: '/usr/lib/python2.7/dist-packages/instances

This patch mocks out the calls to os.path.exists and fileutils
to ensure that the test is fully isolated from the underlying
filesystem permissions, and that fileutils is called with the
correct values.

Change-Id: I0b137caa90b928ef3c400b21b84221fbc721a33c