Split out profile and contanier init methods
In order to migrate a contianer to another
host the profile must be present on the host. And
the contianer is initialized on the new host based
on the websocket that has been initialized on the
source host.

Since there is no real difference in migration,
both cold and live migration, we can re-use the code
for both.

Change-Id: I671b858fa6ba733ccb49a48ae51714cf0c1b0961
Signed-off-by: Chuck Short <chuck.short@canonical.com>