Fix traceback when migrating containers
When live-migrating an instance from one host to
another, an exception was being raised when the host
was being validated. The object that was being checked
via an RPC call was returning a NoneType.

This was due to a typo in returning the value of the
RPC object when the host check was being preformed.

Change-Id: I0d4a5572f84d5978cd0cb8816bee04f894f946f2
Signed-off-by: Chuck Short <chuck.short@canonical.com>