Fix source and destination host
Fix the typo when creating containers
that are migrated across hosts. If we
do not specify the hosts that the container
is to be initialized, nova-lxd will
always try to start the container on the same
host thus throwing an exception.

Add unit test to verify an exception in
case the container initialization fails.

Change-Id: Iab0538599bf7d0d819947f9b14979ce92cbbe675
Signed-off-by: Chuck Short <chuck.short@canonical.com>