Add support for ephemeral storage
Add previously missing ephemeral storage support for nova-lxd.
This patch provides the infrastructure to translate the configuration
of the instance to something that LXD can understand and
act upon.

The way that the ephemeral storage works is that block_device_info dict
and instance object. If an empeheral storage device is needed
then the storage deivce is created on the compute host and bind
mounted into the container when the container is started. This
is due to the contanier is running unprivileged and restricted
mount(8)

To come is restricting the ephemeral storage based on the
LXD storage backend on the compute host.

Update exist unit tests and adds unit new tests.

Change-Id: I923e9eef4caf042a1b11294a4a82bf65a273902e
Signed-off-by: Chuck Short <chuck.short@canonical.com>