Modernize more of LXDDriver
This includes adding an explicit cleanup_host (because it's important
to acknowledge explicitly nothing is needed for cleanup), get_info
using the new pylxd api, and more tests for list_instances.

There were a number of methods that didn't need to be implemented by
the lxd driver as they were identical to the parent class ComputeDriver.

I also explicitly removed the "catch/log/raise" exception patterns,
as I'd rather explicitly catch the exceptions and move on, or allow
nova to handle the exception itself (and log it).

Change-Id: I8c14da579d51c532e279d95d737b90580a790c03