Add support for persistent block devices
This patch adds support for persistent block
deivces via ISCSI.

The way the this works is when the block device is attached via
cinder using os-brick. The block device is attached to the
container and the apparmor profile is updated. Then the
profile is saved.

When the block device is removed the device is remove,
removed vis os-brck and the profile is saved.

Change-Id: I8bd0da919b616c11ac8da4b1a45a8d7ba1a7d907
Signed-off-by: Chuck Short <chuck.short@canonical.com>