Recreate instance bridges after reboot
Currently, nova-lxd instances cannot be restarted
after the host is rebooted.  This is because there
is no logic to created the needed bridges to connect
to the instances.  This leaves instances in an
unbootable state without manual administrator intervention.

This change ensures the network is plugged and the proper
firewall settings are applied to each instance.

Closes-Bug: #1611500

Change-Id: I6fdd321dcdc93124580bab3a90e0254290019a79