Check for container exists before deletion
When a user creates an instance, the user is able to destroy
the instance before LXD has a chance to start the container.
This results in a bad state between LXD and nova-lxd and
pylxd raising an exception that the container was not found.

Before destroying the conainer, check the status code
of LXD to see if the container exists. If the container
does not exist just cleanup the profile if it exists
and the assoicated vifs if had already been created.

This was discovered by running tempest unit tests.

Change-Id: I7c244e8bca9bab69ead5d45c966674e9e7107f20
Signed-off-by: Chuck Short <chuck.short@canonical.com>