Dont raise exception if instance is not found
If the instnace has been deleted before it could be created
pylxd will raise an exception if the container is not found.

This is the correct thing to do in my opinon. However in
nova when the exception is raised the instance will be put
into an error state, since instance life cycle did
not complete.

This was discovered by tempest unit tests.

Change-Id: Ia59cf13095dbbe933c5b96b4adb3f64a8dbe37ca
Signed-off-by: Chuck Short <chuck.short@canonical.com>