Fix usuability issue when attaching block device
When attaching a block device to the container, nova's
client allows the user to chose the block device to be
used. When nova-lxd, the block device that is created
inside the container has the same device minor/major
as the host. This can cause confusion for the user when
the user attaches a block device to the instance and
the block device in the container does not match reality.

To solve this, we create the device inside the container
as the user has requested. However, we specifity the device
major/minor of the host's block device in the LXD profile
so that the container can access the block device
that was made available to the container via cinder.

This was discovered by tempest while running unit tests
against nova-lxd.

Change-Id: I3be43168fe39efeeef5f7b6e57a45e93e3738820
Signed-off-by: Chuck Short <chuck.short@canonical.com>