Add LXC cirros image to devstack configuration
Running a full tempest against nova-lxd with the Ubuntu image
can cause some tests to fail when they would otherwise pass
due to the number of tests running at a given time.

To mitigate we install the cirros LXD image a lighter
image to run the tempest tests against nova-lxd. This is
what upstream uses and allows us to use some of the default
tempest configuration used by devstack.

Change-Id: I6fd58edd49b4e3f7e324be7aaf074db379998af6
Signed-off-by: Chuck Short <chuck.short@canonical.com>