Reset vifs after reboot
When a host is rebooted instances are set into a shutdown instance
state, which is the correct thing.

However when a user tries to restart an instance, LXD will fail
to restart the container because the bridge is not present.

To correct this we re-add the the plug_vifs and unplug_vifs
method.

LP: #1630891

Change-Id: Iae80f4667debab4d1e21d67d882ed52ef59cf15b
Signed-off-by: Chuck Short <chuck.short@canonical.com>