Wait for contianer to start before querying state
When spawning a container nova-lxd does not wait
for the container to spawn before querying the power
state. This results in false feedback to the user
when running 'nova list'.

To fix this use container.start(wait=True) to tell
the LXD API to wait for the instance to change its state.
If the container fails to boot the container will timeout
after 30 seconds.

Closes-Bug: #1636706

Change-Id: Ia089c6430af4d4aee4b3a9d319bc2ff1cfbe39a8
Signed-off-by: Chuck Short <chuck.short@canonical.com>