Remove the need for the `utils` module
utils had the issue that many developing python libraries get:
they develop a "junk drawer" module, usually called "utils". In this
case, most of it is not needed, because we're adding onto a stack
without a re-use case and adding a level of indirection. In at least
one case, a variable substituted for a function call.

I created an adapter that takes an instance and returns
InstanceAttributes that has the reusabled paths (and is also immutable).
In the cases where things weren't getting the benefit of re-use (i.e.
single use functions), I inlined the functionality.

Change-Id: Iac2a6475e10bc403f0e3c693c110f143f083042d