Fix permissions on the config drive
When attachnig a directory it must have the same permissions for
the configdrive so that the user can access files. In this case,
the configdrive permissions were not correct so the user was
not able to ssh into the instances with their ssh keys.

This was already applied to the xenial package. The backport
wasn't completed.

Change-Id: Ia0fb5515f1691bbaa250efdb82195617b107d332