Remove some unneeded/unused dependencies
Dependencies in openstack are sometimes a little hairy (resulting
in some painful mismatches) and the Openstack requirements bot
only complicates things. So I removed the ones we aren't actually
using.

There's also some removal of configs for internationalization that
haven't worked since right after xenial was released, and that we
aren't using and don't really have a need for.

Change-Id: If5c1725cf420174cd75cbd93e1915e276adfe46e