Move nova-lxd to use os-vif.
This patch introduces a new vif driver that implements only plug/unplug.
Those methods merely act as a thin layer over the os-vif functionality,
which requires some adaptation between nova instances/vifs and os-vif
instances/vifs.

The rest of the functionality that was once part of the vif driver has
been moved to functions, as there was no need to make them member methods
(i.e. they never needed `self`). Rather than using the dynamic function
call via string interpolation pattern (which was taken from the libvirt
driver, but is still fragile). Instead, I opted for an explicit map
for config generation.

This patch has 100% test coverage.

The vif library was in dire need of attention, and as a result of the
os-vif transition, it got it.

Change-Id: I175f4df9b58b038b594a40c16076b06f2521bcfc