Consolidate profile creation tooling
`create_profile` actually spanned a number of functions. This patch
smooshes most of it together in a way that makes sense without also
having a 500 line method.

A side effect of this work is that a number of testing hacks can
be removed and we get better unit tests for LXDDriver.

Change-Id: Iec75e84ae247b403dc66a530c7ca44c6e757d6ee