Refactor the flavor <-> profile work
There is a single interface to pass in an instance and get back
a matching profile for the instance now: `flavor.to_profile`. It
uses a map of functions to generate the config and devices from
the instance and network/block device info.

All the unit tests that are added here were added *before* the
beginning of the refactor, so there should be no broken functionality
in this refactor. Along with that, this increases our test coverage
by about 15%.

Change-Id: I575054422386462df170a0a75fac74fbec6f6086