Handle tap device plug/unplug events for linuxbridge
Neutron recently changed behaviour to complete all bridge setup
and configuration of the linuxbridge driver via the neutron agent,
requiring Nova simply to setup the tap device that subsequently
gets plugged into a bridge.

Rework plug/unplug handling based on libvirt driver to fallback
to a legacy style plug/unplug driver for tap devices of this kind,
which in the case of LXD defer the creation of the device to LXD
later in the instance creation process.

Change-Id: I2fdf41e5640f5ca5e3bcd7df1aa159a65b706138
Closes-Bug: 1694719