Handle tap device plug/unplug events for linuxbridge
Neutron recently changed behaviour to complete all bridge setup
and configuration of the linuxbridge driver via the neutron agent,
requiring Nova simply to setup the tap device that subsequently
gets plugged into a bridge.

Rework plug/unplug handling based on libvirt driver to fallback
to a legacy style plug/unplug driver for tap devices of this kind,
which in the case of LXD we actually still use a veth pair so that
a) security group rules are applied to the tap named device on the
host and b) the container still gets part of a nic to use, named
ethXX.

Rework LXD device profile naming for consistency; VIF's attaching
to bridges will be named inline with the bridge, VIF's not being
attached to a bridge will be named with the VIF devname.

Change-Id: I2fdf41e5640f5ca5e3bcd7df1aa159a65b706138
Closes-Bug: 1694719