Re-check image alias prior to sync from glance
If two processes attempt to upload the same image at the same
time, the second will fail as the image alias will already have
been created by the first process.

Re-check by image alias as part of the lock context for the image
prior to syncing the image from glance to ensure that a duplicate
sync is not executed.

Change-Id: Ie5ef1fcf9170e18cf41767683036162fd509ee94
Closes-bug: 1697455