Tidy tox targets, fix coverage reporting
Misc fixes to bring tox targets back inline with general OpenStack
standards (py27 and py35 targets executed by default).

Deal with running 27 and 35 tests in succession.

Fixup coverage target, including fixing paths for restructuring
of modules a few cycles back.

NOTE: this also includes changing the internal name of nova-lxd to
nova.virt.lxd (as this appears to be the only way to convince
coverage to collect and report on code coverage stats).

Change-Id: If0079309a4c4b6b57fc6b5e6f0e947926b010f09