Add non-voting nova-lxd functional test
This patch implements a replacement of the legacy
functional test previously implemented using
openstack-ansible. The job and pipeline is defined
in-repo to ensure that the nova-lxd team has complete
control over the voting status, pipelines it is in,
etc.

The test uses the 'func_lxd' tox environment in the
openstack/openstack-ansible-os_nova repository which
executes an integration build of nova services in a
container, using the nova-lxd compute driver, using
linuxbridge for networking and using the nodepool host
for compute.

Depends-On: If39e4bec8db0be04b13dfc005ee404b73c0b65d3
Change-Id: I49075be82daf62b1ca564c77bf835ae83cd6595b