Fix Py2/Py3 difference in '/' causing py3 unit test failure
Py2 and Py3 treat '/' differently, but '//' the same.  The
semantics of the change mean that the py3 '//' division is required
(i.e. resulting in an integer) for the unit tests to pass

Also includes a backport from master of devstack/tempest-dsvm-lxd-rc
to disable new tests that stable/ocata doesn't/won't support.

Include partial change to address deleting rescued containers during
devstack-gate tests.

Change-Id: Icd83ee747d7287b1cd0c614b3928084262fdb16d