Collection of fixes to unstick the branch with the gate
1. Fix Py2/Py3 difference in '/' causing py3 unit test failure

Py2 and Py3 treat '/' differently, but '//' the same.  The
semantics of the change mean that the py3 '//' division is required
(i.e. resulting in an integer) for the unit tests to pass

2. backport from master of devstack/tempest-dsvm-lxd-rc

Also includes a backport from master of devstack/tempest-dsvm-lxd-rc
to disable new tests that stable/ocata doesn't/won't support.

3. cherry-pick: 6b56b30e8b

Also includes a cherry-pick from 6b56b30e8b which adds a test for
destroying an instance when in rescue mode, and cleans up the code
around deletion of the rescue container.

Change-Id: Icd83ee747d7287b1cd0c614b3928084262fdb16d