Fix up stable/pike to be able to pass tempest tests
1. Backport tempest-dsvm-lxd-rc from master to stable/pike

This is just to ignore some newer/changed tempest tests that the
pike release of nova-lxd doesn't support.

2. Fix up destroy method to ensure proper clean-up

Cherry-picked from: 6b56b30e8b.  This cleans up the destroy method
to ensure that both the containers are deleted if the destroy method is
called whilst and instance is in rescued mode.

Change-Id: Iaceb4bd9c91c07471259cea8e3e337150823c702