Add a test for destroying an instance when in rescue mode
A previous commit added some code to delete the rescued comtainer
when the instance was in rescue mode.  This means that there are two
containers for the instance (and thus using the profile).  However, no
test was added at that stage.  This patchset adds a test to verify that
the containers do get deleted and modifies the destroy code to also
ensure that the rescue container is also stopped before deletion.

Cherry-picked from: 6b56b30e8b56568b88638384cd215692db9403ff

Change-Id: I7f81167c098dfdc24735f2f841cd21f6ce2cb1f4