Enable ZFS to work with datasets and all locales
This fixes two bugs (independently) to get nova-lxd to work with
datasets (pool/set) as well as pools (just 'pool').  It also, switches
the attributes to use the '-p' option with 'zfs list' and 'zpool list'
which returns the values in bytes rather than human readable format,
which is more useful in nova-lxd.

Change-Id: I2eb1b3ba7bd482e680933808414a3992a9a1feba
Closes-Bug: #1756900
Closes-Bug: #1757371