'create_ovs_vif_port' function moved to os-vif
A few functions from nova migrated to os-vif, and two were the
create_ovs_vif_port and delete_ovs_vif_port that nova-lxd uses to
do the post vif wiring on the container to connect it to the appropriate
bridge.  This just gets the correct import, and adds a bit to the
docstrings to help maintainers understand the code.

Change-Id: I0787a919275c544e4faddc9d42d8db5880ba57ee
Closes-Bug: #1776205