Fix bug in scenario/manager and cover it with pep8
manager.py had a syntax error which caused the module to be not
installable.  This wasn't picked up in testing.  This fixes the bug
and adds pep8 coverage to the nova_lxd_tempest_plugin so that the module
always remains installable.

Also, this fixes the tempest concurrency at 1.  See the related bug

Related-Bug: #1790943
Change-Id: Ie542b062a9f0997f6363877ff03f0f9bfc8f3d43