Add a feature support matrix
This patch adds a feature support matrix of nova-lxd.
Items and descriptions of matrix are basically quoted from
the support matrix of nova:
https://docs.openstack.org/nova/latest/user/support-matrix.html

"Block storage over RBD(Ceph)" and "VXLAN networking" are
added newly.

The nova's matrix contains items other than operation.*.
(e.g. networking.*, storage.*, console.*...)
But, those items are not appeared in the built html.
This is a spec of sphinx_feature_classification.support_matrix:
https://docs.openstack.org/sphinx-feature-classification/latest/user/index.html
So the matrix added by this patch changed all items to operation.*.

Change-Id: Iafb072b32f0d0568e04864afef8e053ab442f41a