Fix a tox environment name from func_lxd to lxd
The openstack-ansible-nova-lxd test of Zuul check is failed now.
This test uses a tox environment named as "func_lxd".
The "func_lxd" was defined in the
openstack/openstack-ansible-os_nova repository.

But, the "func_lxd" was renamed to "lxd" in the following patch:
https://review.openstack.org/#/c/600838/

So this patch changes the environment name from "func_lxd" to "lxd".

Change-Id: Idbb12a07fe5dae71b603d5cb6f238bf2bee89a5f