Fix sizing and quotas nova to lxd on unsupported backends
The dir and lvm backends don't support sizing and quotas.  This was made
more complicated with LXD3 as "storage" became the default and the dir
backend, if used, has to be set up as a storage pool.  LXD3 is the
default on bionic.  To resolve an openstack-ansible issue and lay the
ground work for nova-lxd and storage pools and the lxd charm, this patch
makes resolves sending sizes and quotas for a profile root device on lvm
and dir backends.

Change-Id: Idba4f5f3c6015616160d85f25a658cb7e5c43652
Closes-Bug: #1698761