fix error on delete interface when destroying instance
An instance may be destroyed prior nova-compute receives a
vif-delete-event. detach_interface() will so try to get the lxd
profile related to the instance which does not exist anymore, the
process will fail with a NotFound exception raised by lxclib.

In this commit will solves the issue by handle the exception and pass
on updatating the lxc profile. We still continue the process of
vif_driver.un_plug() to ensure all got cleaned.

Change-Id: If7fd5dd17323ec625df3518311dfb012302b2711
Closes-Bug: #1808819
Signed-off-by: Sahid Orentino Ferdjaoui <sahid.ferdjaoui@canonical.com>