Move calling pylxd.client after Tempest tests initialized
This commit moves calling pylxd.client.Client into resource_setup(). If
we have this in __init__(), the test list command `tempest run -l` is
failed without lxd client settings. It's not desirable. And this commit
also renamed variables for lxd client in the code to avoid confusion
with the other clients.

Change-Id: I5498789b2b1e4be6727b6520fc80ce3bace25cf3