On Bionic ensure that a default storage backend is configured
Sets up the default dir backend and configures it as the default pool
for lxd.  This is required on bionic (LXD 3.0.x) as there is no default
storage pool by default.

Also installs apparmor which LXD requires to operate, but that is
missing from ubuntu image that is built for devstack-gate.

Related-Bug: 1822182
Change-Id: I8cc83486d2e040e02dfbeec570d8c8b59b2c2a97