devstack: remove bionic condition when init lxd storage
This make devstack working on all ubuntu based version that is shipped
with LXD3. For versions that ship LXD2 this do not cause any
incidence.

Note that: When using LXD_POOL_NAME with dir backend. LXD2 does not
have the concept of pools so it can't be anything else than "default",
for LXD3 this can be updated but LXD should have been initialized and
the pool created.

Change-Id: I781fb97100937e5189994f3752d0b95b15614d85
Signed-off-by: Sahid Orentino Ferdjaoui <sahid.ferdjaoui@canonical.com>